requirements.txt
absl-py==1.3.0
brotlipy==0.7.0
certifi
cffi
charset-normalizer
cloudpickle==2.2.0
contourpy==1.0.6
cryptography
cycler==0.11.0
Cython==0.29.32
fasteners==0.18
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
glfw==2.5.5
gym==0.26.2
gym-notices==0.0.8
idna
imageio==2.22.3
importlib-metadata==5.0.0
kiwisolver==1.4.4
matplotlib==3.6.2
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
mujoco-py==2.1.2.14
numpy==1.23.4
packaging==21.3
patchelf==0.16.1.0
Pillow==9.3.0
protobuf==3.20.1
pycparser
pyOpenSSL
pyparsing==3.0.9
PySocks
python-dateutil==2.8.2
requests
six
smmap==5.0.0
tensorboardX==2.5.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
typing_extensions
urllib3
zipp==3.10.0
torchinfo
wandb